pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "extra-mymdc"
authors = [
    {name = "James Wrigley", email = "james.wrigley@xfel.eu"}
]
license = {file = "LICENSE"}
requires-python = ">=3.9"
classifiers = ["License :: OSI Approved :: BSD License"]
dynamic = ["version", "description"]
readme = "README.md"
dependencies = [
    "pyyaml",
    "requests",
    "matplotlib"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov"
]

[project.urls]
Home = "https://github.com/European-XFEL/extra-mymdc"
